cognitive_brain/requirements.txt
# Cognitive Brain Requirements

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Deep Learning
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# LLM Inference
vllm>=0.2.6
llama-cpp-python>=0.2.20

# Vector Store & RAG
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Agent Orchestration
langgraph>=0.0.40
langchain-core>=0.1.0

# Async & HTTP
aiohttp>=3.9.0
httpx>=0.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Optional: Type Checking
beartype>=0.16.0
